requirements.txt
arrow==1.3.0
attrs==25.3.0
boto3==1.37.38
botocore==1.37.38
bravado==11.1.0
bravado-core==6.1.1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.1.1
cycler==0.12.1
filelock==3.16.1
fonttools==4.57.0
fqdn==1.5.1
fsspec==2025.3.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
importlib_resources==6.4.5
isoduration==20.11.0
Jinja2==3.1.6
jmespath==1.0.1
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kaleido==0.2.1
kiwisolver==1.4.7
Levenshtein==0.25.1
MarkupSafe==2.1.5
matplotlib==3.7.5
monotonic==1.6
mpmath==1.3.0
msgpack==1.1.0
MuseParse==1.1.9
narwhals==1.37.1
neptune==1.14.0
networkx==3.1
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.11.0.86
packaging==25.0
pandas==2.0.3
pillow==10.4.0
pkgutil_resolve_name==1.3.10
plotly==6.0.1
poppler-utils==0.1.0
psutil==7.0.0
py-cpuinfo==9.0.0
PyAudio==0.2.14
pydub==0.25.1
pyfluidsynth==1.3.4
PyJWT==2.9.0
PyMusic-Instrument==0.5
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
rapidfuzz==3.9.7
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.1
s3transfer==0.11.5
scipy==1.10.1
seaborn==0.13.2
simpleaudio==1.0.4
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
swagger-spec-validator==3.0.4
sympy==1.13.3
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.130
ultralytics-thop==2.0.14
uri-template==1.3.0
urllib3==1.26.20
utils==1.0.2
webcolors==24.8.0
websocket-client==1.8.0
zipp==3.20.2
